+++ CHANGELOG.md
@@ -20,6 +20,10 @@ Please see the [DBUG issue queue](https://github.com/TransferWare/dbug/issues) f
 
 ## [Unreleased]
 
+### Added
+
+- [Create channel for Logger.](https://github.com/TransferWare/dbug/issues/19)
+
 ## [7.3.0] - 2025-12-13
 
 ### Fixed